src/requirements.txt
fhir.resources==6.5.0
pytest==7.3
psycopg[binary]==3.1.8
